+++ src/main/java/kr/hhplus/be/domain/order/entity/Order.java
@@ -4,7 +4,7 @@ import jakarta.persistence.*;
 import kr.hhplus.be.domain.order.enumtype.OrderStatus;
 import lombok.*;
 
-import java.time.LocalDate;
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Getter
@@ -24,11 +24,11 @@ public class Order {
 
     private LocalDateTime orderedAt;
 
-    private int totalAmount;
+    private BigDecimal totalAmount;
 
-    private int discountAmount;
+    private BigDecimal discountAmount;
 
-    private int finalAmount;
+    private BigDecimal finalAmount;
 
     @Enumerated(EnumType.STRING)
     private OrderStatus status;
