+++ src/main/java/kr/hhplus/be/domain/product/dto/TopSalesProductDTO.java
@@ -6,6 +6,8 @@ import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Getter
 @AllArgsConstructor
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
@@ -13,7 +15,7 @@ public class TopSalesProductDTO {
     private long productId;
     private String productName;
     private Category category;
-    private int price;
+    private BigDecimal price;
     private int soldQuantity;
-    private int totalAmount;
+    private BigDecimal totalAmount;
 }
