+++ src/main/java/kr/hhplus/be/domain/user/dto/BalanceDTO.java
@@ -3,9 +3,11 @@ package kr.hhplus.be.domain.user.dto;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 
+import java.math.BigDecimal;
+
 @Getter
 @AllArgsConstructor
 public class BalanceDTO {
     long userId;
-    int balance;
+    BigDecimal balance;
 }
