+++ src/main/java/kr/hhplus/be/interfaces/user/dto/ChargeRequest.java
@@ -5,9 +5,11 @@ import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 public class ChargeRequest {
     @NotNull(message = "충전금액은 필수값 입니다.")
-    private Integer chargeAmount;
+    private BigDecimal chargeAmount;
 }
