+++ src/test/java/kr/hhplus/be/fixture/ProductFixture.java
@@ -4,11 +4,12 @@ import kr.hhplus.be.domain.product.enumtype.Category;
 import kr.hhplus.be.domain.product.entity.Product;
 import kr.hhplus.be.domain.product.entity.ProductInventory;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 public class ProductFixture {
 
-    public static Product createProduct(String name, Category category, int price) {
+    public static Product createProduct(String name, Category category, BigDecimal price) {
         return new Product(null, name, category, price, LocalDateTime.now(), LocalDateTime.now());
     }
